mesa: update 20.2.4 -> 20.3.1
Two of the patches have been applied upstream:
	0001-meson-Add-xcb-fixes-to-loader-when-using-x11-and-dri.patch
	0005-vc4-use-intmax_t-for-formatted-output-of-timespec-me.patch

The 'surfaceless' platform is now enabled by default, therefore it is no
longer a valid choice. The meson build system is intelligent enough now to
decide if 'drm' needs to be enabled based on whether or not gbm is available.
Therefore 'drm' also is no longer a valid platform choice:

	ERROR: Options "drm, surfaceless" are not in allowed choices: "auto, x11, wayland, haiku, android, windows"

Signed-off-by: Trevor Woerner <twoerner@gmail.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>